Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Este proyecto es el funcionamiento por detras de la logica de negocio de unas quinielas"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "reportlab>=4.2.0",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.48",
    "werkzeug>=3.1.6",
]
